Fix the clang-tidy findings clang-tidy cannot
The remaining findings, by hand, taking check-clang-tidy from 1083 findings to none on both
macOS and Linux.

Real defects and dead code:
  * A vestigial ternary in the TypeScript emitter whose branches were both "  | ".
  * A member of Server that nothing read once its only user became static.
  * Unused locals in the Rust and TypeScript emitters, a pointless std::move, and a conditional
    chain in the uavcan catalogue that assigned maxValue from two arms.
  * CEmitter assembled each generated file by concatenating whole buffers, copying the body twice.
  * dsdlc, dsdld and three test tools had no top-level exception handling: anything escaping main
    reached std::terminate with nothing printed. Each now reports and exits non-zero.
  * The unit-test entry points get a real header, UnitTests.h, instead of 59 hand-maintained
    extern declarations in UnitMain.cpp -- which is also what satisfies misc-use-internal-linkage.
  * Rule-of-five gaps on the pimpl and RAII types, and a `language` field left indeterminate.
  * The two branches of environmentWith disagreed on constness. Only the Apple one is compiled on
    macOS, so the Linux spelling had never been looked at.

Two checks are subtracted in .clang-tidy rather than satisfied, with the reason recorded there:
bugprone-unchecked-optional-access cannot model how this tree establishes engagement (MLIR's
FailureOr behind failed(), paired predicates, caller preconditions), and
misc-const-correctness.AllowedTypes now exempts random distributions, whose operator() is not const.

The rest carry NOLINT with the reason at the site: names that mirror std::flat_set and pydsdl,
fixed fuzz seeds that exist to reproduce, a test that asserts moved-from behaviour, getenv reads
that happen before any worker thread starts, MLIR's own idioms, and the emitter concatenations
where an append sequence would cost the reader the line being emitted.

+++ lib/Lowering/LowerToMLIR.cpp
@@ -197,8 +197,8 @@ mlir::OwningOpRef<mlir::ModuleOp> lowerToMLIR(const SemanticModule& module,
                            builder.getStringAttr(cTypeNameFromInfo(def.info, TypeNameVersioning::Unversioned)));
         state.addAttribute("header_path", builder.getStringAttr(relativeHeaderPath(def.info)));
         state.addAttribute("full_name", builder.getStringAttr(def.info.fullName));
-        state.addAttribute("major", builder.getI32IntegerAttr(def.info.majorVersion));
-        state.addAttribute("minor", builder.getI32IntegerAttr(def.info.minorVersion));
+        state.addAttribute("major", builder.getI32IntegerAttr(static_cast<std::int32_t>(def.info.majorVersion)));
+        state.addAttribute("minor", builder.getI32IntegerAttr(static_cast<std::int32_t>(def.info.minorVersion)));
         if (def.request.sealed)
         {
             state.addAttribute("sealed", builder.getUnitAttr());
@@ -227,6 +227,7 @@ mlir::OwningOpRef<mlir::ModuleOp> lowerToMLIR(const SemanticModule& module,
 
         auto* schema     = builder.create(state);
         auto& schemaBody = schema->getRegion(0);
+        // NOLINTNEXTLINE(cppcoreguidelines-owning-memory) -- Region::push_back takes ownership.
         schemaBody.push_back(new mlir::Block());
 
         builder.setInsertionPointToStart(&schemaBody.front());
@@ -331,6 +332,7 @@ mlir::OwningOpRef<mlir::ModuleOp> lowerToMLIR(const SemanticModule& module,
             planState.addRegion();
             auto* plan       = builder.create(planState);
             auto& planRegion = plan->getRegion(0);
+            // NOLINTNEXTLINE(cppcoreguidelines-owning-memory) -- Region::push_back takes ownership.
             planRegion.push_back(new mlir::Block());
 
             builder.setInsertionPointToStart(&planRegion.front());
